refactor(inference): move shared suite helpers under inference/utils
Colocate lifecycle, cache probe, and results table modules with other inference utilities and update import paths across IX atom suites and report wiring.
